Update `paramiko` stubs to 3.0.*
