+++ stubs/paramiko/@tests/stubtest_allowlist.txt
@@ -2,11 +2,6 @@ paramiko.SFTPServer.__init__
 paramiko.Transport.open_x11_channel
 paramiko.Transport.send_ignore
 paramiko.Transport.start_server
-paramiko.py3compat.input
-paramiko.py3compat.BytesIO.readlines
-paramiko.py3compat.BytesIO.seek
-paramiko.py3compat.StringIO.seek
-paramiko.py3compat.StringIO.truncate
 paramiko.sftp_server.SFTPServer.__init__
 paramiko.transport.Transport.open_x11_channel
 paramiko.transport.Transport.send_ignore
