+++ stubs/paramiko/METADATA.toml
@@ -1,4 +1,4 @@
-version = "2.12.*"
+version = "3.0.*"
 # Requires a version of cryptography where cryptography.hazmat.primitives.ciphers.Cipher is generic
 requires = ["cryptography>=37.0.0"]
