+++ stubs/paramiko/paramiko/pkey.pyi
@@ -14,7 +14,7 @@ class PKey:
     END_TAG: Pattern[str]
     def __init__(self, msg: Message | None = ..., data: str | None = ...) -> None: ...
     def asbytes(self) -> bytes: ...
-    def __cmp__(self, other: object) -> int: ...
+    def __bytes__(self) -> bytes: ...
     def __eq__(self, other: object) -> bool: ...
     def get_name(self) -> str: ...
     def get_bits(self) -> int: ...
