+++ stubs/paramiko/paramiko/util.pyi
@@ -1,9 +1,8 @@
 from _typeshed import Self
-from collections.abc import Callable
 from hashlib import _Hash
 from logging import Logger, LogRecord
 from types import TracebackType
-from typing import IO, AnyStr, Protocol, TypeVar
+from typing import IO, AnyStr, Protocol
 
 from paramiko.config import SSHConfig, SSHConfigDict
 from paramiko.hostkeys import HostKeys
@@ -11,13 +10,7 @@ from paramiko.hostkeys import HostKeys
 class SupportsClose(Protocol):
     def close(self) -> None: ...
 
-_T = TypeVar("_T")
-
 def inflate_long(s: bytes, always_positive: bool = ...) -> int: ...
-
-deflate_zero: int
-deflate_ff: int
-
 def deflate_long(n: int, add_sign_padding: bool = ...) -> bytes: ...
 def format_binary(data: bytes, prefix: str = ...) -> list[str]: ...
 def format_binary_line(data: bytes) -> str: ...
@@ -36,7 +29,6 @@ class PFilter:
     def filter(self, record: LogRecord) -> bool: ...
 
 def get_logger(name: str) -> Logger: ...
-def retry_on_signal(function: Callable[[], _T]) -> _T: ...
 def constant_time_bytes_eq(a: AnyStr, b: AnyStr) -> bool: ...
 
 class ClosingContextManager:
@@ -46,3 +38,9 @@ class ClosingContextManager:
     ) -> None: ...
 
 def clamp_value(minimum: int, val: int, maximum: int) -> int: ...
+
+# This function attempts to convert objects to bytes,
+# *but* just returns the object unchanged if that was unsuccessful!
+def asbytes(s: object) -> object: ...
+def b(s: str | bytes, encoding: str = "utf-8") -> bytes: ...
+def u(s: str | bytes, encoding: str = "utf-8") -> str: ...
